rename playerEdit & teamEdit to update Team/Player respectively
renamed to match names accordinly to functionality
