+++ template-engine/public/src/team-editor/updatePlayer.js
