+++ template-engine/public/src/team-editor/updateTeam.js
